Time: 1 ms (99.75%) | Memory: 43.3 MB (71.56%) - LeetSync

+++ 3379-score-of-a-string/score-of-a-string.java
@@ -3,7 +3,7 @@ class Solution {
         int sum=0;
         for(int i=0;i<s.length()-1;i++)
         {
-            sum+=Math.abs((int)s.charAt(i)-(int)s.charAt(i+1));
+            sum+=Math.abs(s.charAt(i)-s.charAt(i+1));
 
         }
         return sum;
